Changes in Version 0.1.3
Fixed main animation interval errors. When animations finished, the loop
called play() on them which caused them to be removed from the
Animation.animations array which caused the for-loop to exceed the
array's length. To fix this issue, I simply used a copy of the array
rather than just referencing it, and used that copy for the for-loop
during the interval. This way the copy will remain unchanged during the
for-loop, while the original can effectively be modified.