Changes in 0.1.4
Added support for an options parameter for addFrame. Current options are begin, step, end, duration and extra. Begin and end callback are not implemented yet and do not work.

Renamed frame callback functions from 'callback' to 'step'.

Used Math.min and Math.max in scrub() to keep 'time' within the range of 0 to the animation duration.

Added an event.html test page for testing event callbacks in the future.